requirements.txt
appdirs==1.4.4
apturl==0.5.2
astroid==2.4.2
autopep8==1.5.4
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
CacheControl==0.12.6
cachelib==0.1.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fake-useragent==0.1.11
fasteners==0.14.1
filelock==3.0.12
firebase-admin==4.4.0
Flask==1.1.2
Flask-Session==0.3.2
future==0.18.2
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.5.0
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parse==1.18.0
pbr==5.5.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==2020.8.13
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
pyee==7.0.4
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyppeteer==0.2.2
pyquery==1.4.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==1.5
python-debian===0.1.36ubuntu1
python-dotenv==0.14.0
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-html==0.10.0
requests-unixsocket==0.2.0
rsa==4.6
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
stevedore==3.2.2
systemd-python==234
toml==0.10.1
tqdm==4.50.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.10
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
w3lib==1.22.0
wadllib==1.3.3
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
